Move batched behavior into subclassed worker
