+++ lib/chore/strategies/worker/single_batched_worker_strategy.rb
@@ -1,0 +1,16 @@
+require 'chore/strategies/worker/single_worker_strategy'
+
+module Chore
+  module Strategy
+    class SingleBatchedWorkerStrategy < SingleWorkerStrategy
+      def assign(work)
+        if workers_available?
+          @worker = BatchedWorker.new(work, @options)
+          @worker.start
+          @worker = nil
+          true
+        end
+      end
+    end
+  end
+end
